wrapper to malloc in main

+++ seta/msg.c
@@ -26,7 +26,7 @@ msg_t msg_new_from_str(char *str) {
 msg_t msg_new_from_int(int i) {
 	char aux[30];
 	sprintf(aux, "%d", i);
-	int dim_new = sizeof(char) * (strlen(aux) + 1);
+	int dim_new = sizeof(char) * ((int)strlen(aux) + 1);
 	char *new_str = (char *)malloc(dim_new);
 	memset(new_str, '\0', dim_new);
 	strncpy(new_str, aux, dim_new - 1);
